Move separating planes to ConvexSurfaceMesh2d
