+++ hoomd-geometry/src/shape/convex_polytope.rs
